+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Added ###
 
   * Built-in filter bags can now be disabled
+  * New filter bag for Anima items
 
 ### Changed ###